Clean up a few compile warnings

+++ src/engine/client/text.cpp
@@ -299,7 +299,7 @@ class CTextRender : public IEngineTextRender
 		int SlotSize = SlotW*SlotH;
 		int x = 1;
 		int y = 1;
-		int px, py;
+		unsigned int px, py;
 
 		FT_Set_Pixel_Sizes(pFont->m_FtFace, 0, pSizeData->m_FontSize);
 
@@ -340,8 +340,8 @@ class CTextRender : public IEngineTextRender
 				}
 		}
 
-		if(0) for(py = 0; py < SlotW; py++)
-			for(px = 0; px < SlotH; px++)
+		if(0) for(py = 0; (int)py < SlotW; py++)
+			for(px = 0; (int)px < SlotH; px++)
 				ms_aGlyphData[py*SlotW+px] = 255;
 
 		// upload the glyph
